fix(format): use portable format specifiers for int64_t/uint64_t
On LP64 (64-bit Linux/macOS), int64_t is long (8 bytes), not long long.
Using %lld for int64_t arguments causes -Wformat warnings because the
format specifier expects unsigned long long but the argument type is long.

Fix by casting int64_t arguments to (long long) and uint64_t to
(unsigned long long) at format string call sites. Also fix:
- PDSCLIDispatcher: cast data.bytes to (const char *) for %s
- WebSocketServer: use gai_strerror(gai_err) instead of raw int for %s

+++ Garazyk/Sources/PLC/PLCDIDKey.m
@@ -125,7 +125,7 @@ static NSUInteger PLCReadVarint(const uint8_t *bytes, NSUInteger maxLength, uint
         if (error) {
             *error = [NSError errorWithDomain:PLCDIDKeyErrorDomain
                                          code:6
-                                     userInfo:@{NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unsupported multicodec: 0x%llx", multicodec]}];
+                                     userInfo:@{NSLocalizedDescriptionKey: [NSString stringWithFormat:@"Unsupported multicodec: 0x%llx", (unsigned long long)multicodec]}];
         }
         return nil;
     }
